Fixed install failures

+++ functions/classes/class.Install.php
@@ -44,10 +44,8 @@ class Install extends Common_functions {
 		# Log object
 		try { $this->Database->connect();
 		} catch ( Exception $e ) {
-
+			$err = true;
 		}
-		if(!isset($e))
-		$this->Log = new Logging ($this->Database);
 	}
 
 
@@ -99,6 +97,7 @@ class Install extends Common_functions {
 
 	    # return true, if some errors occured script already died! */
 		sleep(1);
+		$this->Log = new Logging ($this->Database);
 		$this->Log->write( "Database installation", "Database installed successfully. Version ".VERSION.".".REVISION." installed", 1 );
 		return true;
 	}
@@ -156,22 +155,25 @@ class Install extends Common_functions {
 	    $query  = file_get_contents("../../db/SCHEMA.sql");
 
 	    # formulate queries
-	    $queries = explode(";\n", $query);
+	    $queries = array_filter(explode(";\n", $query));
 
 	    # execute
 	    foreach($queries as $q) {
-			try { $this->Database_root->runQuery($q.";"); }
-			catch (Exception $e) {
-				//unlock tables
-				$this->Database_root->runQuery("UNLOCK TABLES;");
-				//drop database
-				try { $this->Database_root->runQuery("drop database if exists ". $this->db['name'] .";"); }
+		    //length check
+		    if (strlen($q)>0) {
+				try { $this->Database_root->runQuery($q.";"); }
 				catch (Exception $e) {
-					$this->Result->show("danger", 'Cannot set permissions for user '. $db['user'] .': '.$e->getMessage(), true);
+					//unlock tables
+					$this->Database_root->runQuery("UNLOCK TABLES;");
+					//drop database
+					try { $this->Database_root->runQuery("drop database if exists ". $this->db['name'] .";"); }
+					catch (Exception $e) {
+						$this->Result->show("danger", 'Cannot set permissions for user '. $db['user'] .': '.$e->getMessage(), true);
+					}
+					//print error
+					$this->Result->show("danger", "Cannot install sql SCHEMA file: ".$e->getMessage()."<br>query that failed: <pre>$q</pre>", false);
+					$this->Result->show("info", "Database dropped", false);
 				}
-				//print error
-				$this->Result->show("danger", "Cannot install sql SCHEMA file: ".$e->getMessage()."<br>query that failed: <pre>$q</pre>", false);
-				$this->Result->show("info", "Database dropped", false);
 			}
 	    }
 	}
@@ -369,6 +371,7 @@ class Install extends Common_functions {
 	    foreach($queries as $query) {
 			try { $this->Database->runQuery($query); }
 			catch (Exception $e) {
+				$this->Log = new Logging ($this->Database);
 				# write log
 				$this->Log->write( "Database upgrade", $e->getMessage()."<br>query: ".$query, 2 );
 				# fail
@@ -379,6 +382,7 @@ class Install extends Common_functions {
 
 		# all good, print it
 		sleep(1);
+		$this->Log = new Logging ($this->Database);
 		$this->Log->write( "Database upgrade", "Database upgraded from version ".$this->settings->version." to version ".VERSION.".".REVISION, 1 );
 		return true;
 	}
